+++ Vistas/MyUserControl/LoginUserControl.xaml
@@ -1,0 +1,36 @@
+﻿<UserControl x:Class="Vistas.MyUserControl.LoginUserControl"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             mc:Ignorable="d"
+             Background="Transparent"
+             d:DesignHeight="441" d:DesignWidth="322">
+    <UserControl.Resources>
+        <!-- Cargar el ResourceDictionary que contiene los estilos -->
+        <ResourceDictionary Source="../MyStyle.xaml"/>
+    </UserControl.Resources>
+    <Grid Height="404" Width="301">
+        <StackPanel Margin="10,12,12,-12" Orientation="Vertical">
+            <Image Height="100" Source="/images/logofacu.png" Width="100" />
+            <TextBlock FontFamily="Consolas" FontSize="25" FontWeight="Bold" Foreground="White" HorizontalAlignment="Center" Text="GRUPO 3 CODE" />
+            <TextBlock FontFamily="Consolas" FontSize="12" FontWeight="Medium" Foreground="LightGray" Margin="0,3,0,0" Text="Si lo podes imaginar, lo podes programar" TextAlignment="Center" TextWrapping="Wrap" />
+            <TextBlock FontFamily="Consolas" FontSize="12" FontWeight="Medium" Foreground="DarkGray" Margin="0,15,0,0" Text="Usuario" />
+            <TextBox BorderBrush="DarkGray" BorderThickness="0,0,0,1" CaretBrush="LightGray" FontFamily="Consolas" FontSize="13" FontWeight="Medium" Foreground="White" Height="28" Margin="0,5,0,0" Name="txtUser" Padding="20,0,0,0" VerticalContentAlignment="Center">
+                <TextBox.Background>
+                    <ImageBrush AlignmentX="Left" ImageSource="/images/user-icon.png" Stretch="None" />
+                </TextBox.Background>
+            </TextBox>
+            <TextBlock FontFamily="Consolas" FontSize="12" FontWeight="Medium" Foreground="DarkGray" Margin="0,15,0,0" Text="Contraseña" />
+            <PasswordBox BorderBrush="DarkGray" BorderThickness="0,0,0,1" CaretBrush="LightGray" FontFamily="Consolas" FontSize="13" FontWeight="Medium" Foreground="White" Height="28" Margin="0,5,0,0" Name="txtPass" Padding="20,0,0,0" VerticalContentAlignment="Center">
+                <PasswordBox.Background>
+                    <ImageBrush AlignmentX="Left" ImageSource="/images/key-icon.png" Stretch="None" />
+                </PasswordBox.Background>
+            </PasswordBox>
+            <Button Click="btnLogin_Click" Content="Ingresar" Margin="0,50,0,0" x:Name="btnLogin" Style="{StaticResource LoginButtonStyle}" />
+            <StackPanel HorizontalAlignment="Center" Margin="0,15,0,0" Orientation="Horizontal">
+                <TextBlock Cursor="Hand" FontFamily="Consolas" FontSize="12" FontWeight="Medium" Foreground="DarkGray" Text="¿Olvidaste la Contraseña?" />
+            </StackPanel>
+        </StackPanel>
+    </Grid>
+</UserControl>
